Consolidate symbols_linux* files
